feat: 1.21.110 support (#272)
* feat: add optional keyword parameter for `loadLanguageKeywordsFromPack`

* refactor: update connection request structs

* feat: update `ServerNetworkHandler`

* feat: add `Packet::toString`

* refactor: resource pack repository

* feat: update structures

* feat: update network protocol version

* feat: update symbols for BDS 1.21.111.1 (Windows)

* refactor: rename `disconnectClient` to `disconnectClientWithMessage`

* feat: bump Minecraft version to 1.21.111

* fix: correct `Block` class members

* feat: add `canUseAsAttack` and `useAsAttack` to `Item`

* fix: correct virtual tables of `Biome` and `ILevel`

* feat: update symbols for BDS 1.21.111.1 (Linux)

* feat: update classes

* feat: update `Actor` and `Player` class

* fix: resolve various crashes

* fix: correct symbol

* fix: update `LoginPacket`

* fix: server stop hook on Linux

+++ src/bedrock/world/level/block/block.cpp
@@ -42,12 +42,12 @@ Brightness Block::getLight() const
 
 int Block::getFlameOdds() const
 {
-    return direct_data_.flame_odds;
+    return static_cast<int>(direct_data_.flame_odds);
 }
 
 int Block::getBurnOdds() const
 {
-    return direct_data_.burn_odds;
+    return static_cast<int>(direct_data_.burn_odds);
 }
 
 float Block::getExplosionResistance() const
